# docs/swagger.yaml
definitions:
...
  collegemanagementsystem.Student:
    properties:
      age:
        type: integer
      dept:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/students:
    get:
      description: Retrieve all students
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/collegemanagementsystem.Student'
            type: array
      security:
      - BearerAuth: []
      summary: Get all students
      tags:
      - Students
    post:
      consumes:
      - application/json
      description: Add a new student record
      parameters:
      - description: Student Data
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Student'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/collegemanagementsystem.Student'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create new student
      tags:
      - Students
  /api/students/{id}:
    delete:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: student deleted
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete student
      tags:
      - Students
    get:
      parameters:
      - description: Student ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/collegemanagementsystem.Student'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get student by ID
      tags:
      - Students
    put:
      consumes:
      - application/json
      parameters:
      - description: Updated Student
        in: body
        name: student
        required: true
        schema:
          $ref: '#/definitions/collegemanagementsystem.Student'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/collegemanagementsystem.Student'
      security:
      - BearerAuth: []
      summary: Update student
      tags:
      - Students
...
swagger: "2.0"